+++ mes-test-schema/generateSchemas.ts
@@ -47,11 +47,11 @@ if (cliArgument === 'generate') {
     // generateTypescriptInterfaces(combinedSchema);
   }
   // copy Cat A mod 1 and Cat A mod 2 index.json files
-  copyFile(`./category-definitions/AM1/partial.json`, `./categories/AM1/index.json`);
-  copyFile(`./category-definitions/AM2/partial.json`, `./categories/AM2/index.json`);
+  copyFile(`./category-definitions/AM1/index.json`, `./categories/AM1/index.json`);
+  copyFile(`./category-definitions/AM2/index.json`, `./categories/AM2/index.json`);
   // generate Cat A mod 1 and mod 2 typescript files (standalone)
-  generateTypescriptInterfaces(`./category-definitions/AM1/partial.json`, `./categories/AM1/partial.d.ts`);
-  generateTypescriptInterfaces(`./category-definitions/AM2/partial.json`, `./categories/AM2/partial.d.ts`);
+  generateTypescriptInterfaces(`./category-definitions/AM1/index.json`, `./categories/AM1/index.d.ts`);
+  generateTypescriptInterfaces(`./category-definitions/AM2/index.json`, `./categories/AM2/index.d.ts`);
   // copy common index.json file across for consistency
   copyFile(`./category-definitions/common/index.json`, `./categories/common/index.json`)
   console.log('All categories have been generated');
